All low-level kernels moved into fpga tasks
Additional minor changes to makefiles. Changed also the OptimizeProblem
usage. Instead of copying the diagonal values into a new memory space,
now we use an array of integers to save at wich j-position in each i-row
in the global matrix values array is stored the diagonal value. Values
cannot be moved, since TestCG modifies the diagonal of the matrix in
order to test the algorightm.