First version only with dependence-based flow (smp)
Most of the explicit taskwaits have been transformed. However, they cannot
be removed, since hpgc timing method expects functions to start/stop
when they are called (sync), so tick/tock are done around them.

In order to still taking advantage of this method, we have added a
`taskwait noflush` at specific points (at functions that are timed,
basically)

This is not 100% accurate, since, effectively, the tock() call is done
after the task is executed, but the tick() is done before it is
scheduled to execute. This needs to be reviewed.

In order to have both sync and async calls (we need sync calls for the
test/check phases of hpcg), the functions have been renamed as _nw and
wrapped around a new function with the original name that just does the
call to the _nw version and performs a taskwait (with flush)