Blocking + almost-dependences-driven version
This version removes the `pragma omp taskwait`, making it (almost)
dependence-driven. It cannot be 100% dependece-driven because `hpcg`
measures timing execution for each kernel, surrounding the call to each
kernel with a tick/tock macro.

As it is currently written now, this would only measure the task
creation time, not the time execution time. In order to introduce
minimum changes to the CG structure, we've leaved it as it is (with the
tick/tock structure) and added a `omp taskwait noflush` to the kernels.

After each task creation, the main CG thread waits for its execution, so
the tick-tock measuremens are, more or less, accurate.

What we really need to do is be able to replace the tick-tocks
measurements with task-specific measurements. We can leave the calls to
tick-tock there, obtain timings by ourselves and replace the tick-tock
measurements with ours.

Anyway, timings should be taken with pinch of salt right now.