OmpSs@FPGA + MPI working version (test)
Tested in MEEP workstations between two nodes with the Alveo U280.

This version is a branch from the blocking version. Be careful,
dot product is not 100% accurate, it does not work with more than
one accelerator (needs refactoring). However, as proof of concept
(performance is still very poor), it should work (FPGA+MPI)